+++ assets/js/steps/components/Banner.jsx
