+++ assets/js/steps/components/Barcode.jsx
