+++ assets/js/steps/components/NavBar.jsx
